+++ src/main/java/com/gang/community/dto/QuestionDTO.java
@@ -5,13 +5,13 @@ import lombok.Data;
 
 @Data
 public class QuestionDTO {
-    private int id;
+    private Long id;
     private String title;
     private String description;
     private String tag;
     private Long gmtCreate;
     private Long gmtModified;
-    private int creator;
+    private Long creator;
     private int commentCount;
     private int viewCount;
     private int likeCount;
